changed group to settings
